+++ lib/src/workers.dart
@@ -3,7 +3,23 @@ import 'dart:core' as core;
 import 'package:web/web.dart' as web;
 import 'package:poor_web/poor_web.dart';
 
+/// Cloudflare Workers API
 extension type Fetcher._(Object _assets) implements Object {
+  /// Fetcher
+  ///
+  /// ```dart
+  /// @JS('__dart_fetch')
+  /// external set fetcher(Fetcher _);
+  ///
+  /// main() => fetcher = Fetcher(_handler);
+  ///
+  /// Future<Response> _handler(Request request, Env env, Context ctx) async =>
+  ///     Response(String('Hello, Dart!'));
+  /// ```
+  factory Fetcher(Handler fetch) => Fetcher._(Object())
+    ..['fetch'] = Fn.a3((_Request request, Env env, Context ctx) =>
+        Promise(fetch(request, env, ctx)));
+
   @JS()
   external Promise<Response> fetch(Request _, Env __, Context ___);
 }